+++ src/Resources/Resources/ChangeLog.md
@@ -24,6 +24,7 @@
 * Introduced secrets detection feature to safeguard sensitive data.
 * Migrated SDK generation from autorest csharp to autorest powershell.
 * Fixed deployment and deployment stack New/Set cmdlets to fail if template/parameter uri fails to downloads.
+* Adjusted Depoyment Stacks cmdlets to prepare for GA.
 
 ## Version 6.16.1
 * Added null check to the permissions object in the ToPSRoleDefinition method.